requirements.txt
ultralytics==8.4.3
opencv-python-headless==4.12.0.88
python-dotenv==1.2.2
requests>=2.31.0
PyYAML>=6.0
onnxruntime>=1.17.0
